Two small suggested edits.  See below for something I feel is a little confusing
In Section 6.2 that describes the Buffer approach there is this sentence:

"The developer then receives ZEP and spends the external tokens."

This is a little confusing to me.  Since the external tokens are described as belonging to the external protocol (Ether?  ZCash?, etc.),  the concept of receiving ZEP tokens and somehow spending the external tokens by spending the ZEP tokens is hard for me to grasp.  Is the intent here to say that the developer, now that they have received ZEP tokens, is spending implicitly spending the external tokens by proxy where the proxy agent is the ZEP tokens?  Without knowing the true intent here I can't suggest an alternate explanation.